refactor: police view mobile audit, auth route reorganization, and component cleanup
