+++ frontend/src/app/(student)/_components/RegistrationStatus.tsx
@@ -2,6 +2,7 @@ import { Card, CardContent } from "@/components/ui/card";
 import { Skeleton } from "@/components/ui/skeleton";
 import { hasActiveHold } from "@/lib/api/location/location.service";
 import { clientEnv } from "@/lib/config/env.client";
+import { getErrorMessage } from "@/lib/errors";
 import { isFromThisSchoolYear } from "@/lib/utils";
 import { format } from "date-fns";
 import { AlertTriangleIcon, CheckCircle, ExternalLink } from "lucide-react";
@@ -25,7 +26,7 @@ export default function RegistrationStatus({
         <CardContent className="p-0">
           <div className="text-center text-destructive py-4">
             <p className="font-semibold mb-1">Error loading course status</p>
-            <p className="text-sm">{error.message}</p>
+            <p className="text-sm">{getErrorMessage(error)}</p>
           </div>
         </CardContent>
       </Card>
