+++ frontend/src/app/(student)/_components/tracker/RegistrationPartyCard.tsx
@@ -9,7 +9,11 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { Skeleton } from "@/components/ui/skeleton";
 import { PartyDto } from "@/lib/api/party/party.types";
-import { formatPhoneNumber, formatTime } from "@/lib/utils";
+import {
+  formatContactPreference,
+  formatPhoneNumber,
+  formatTime,
+} from "@/lib/utils";
 import { format } from "date-fns/format";
 import { Ban, Mail, MoreVertical, Pencil, Phone } from "lucide-react";
 import { memo } from "react";
@@ -94,8 +98,11 @@ const RegistrationPartyCard = memo(function RegistrationPartyCard({
               <p className="flex items-center gap-1.5">
                 <Phone className="size-3 shrink-0" />
                 {formatPhoneNumber(party.contact_one.phone_number)}
-                <span className="capitalize">
-                  - {party.contact_one.contact_preference}
+                <span>
+                  -{" "}
+                  {formatContactPreference(
+                    party.contact_one.contact_preference
+                  )}
                 </span>
               </p>
               <p className="flex items-center gap-1.5">
@@ -114,8 +121,11 @@ const RegistrationPartyCard = memo(function RegistrationPartyCard({
               <p className="flex items-center gap-1.5">
                 <Phone className="size-3 shrink-0" />
                 {formatPhoneNumber(party.contact_two.phone_number)}
-                <span className="capitalize">
-                  - {party.contact_two.contact_preference}
+                <span>
+                  -{" "}
+                  {formatContactPreference(
+                    party.contact_two.contact_preference
+                  )}
                 </span>
               </p>
               <p className="flex items-center gap-1.5">
