+++ frontend/src/app/police/_components/PartyCard.tsx
@@ -24,30 +24,32 @@ import {
   getRemoteWarningCount,
 } from "@/lib/api/location/location.types";
 import { PartyDto } from "@/lib/api/party/party.types";
-import { cn, formatContactPreference, formatTime } from "@/lib/utils";
+import {
+  buildMapsUrl,
+  cn,
+  formatAddress,
+  formatContactPreference,
+  formatTime,
+} from "@/lib/utils";
 import { format } from "date-fns";
-import { AlertTriangle, EllipsisVertical } from "lucide-react";
+import { AlertTriangle, EllipsisVertical, ExternalLink } from "lucide-react";
 import { memo } from "react";
 
 const INCIDENT_MENU_ITEMS: {
   severity: IncidentSeverity;
   label: string;
-  hoverLabel: string;
 }[] = [
   {
     severity: "remote_warning",
     label: `Add ${INCIDENT_SEVERITY_LABELS.remote_warning.toLowerCase()}`,
-    hoverLabel: `${INCIDENT_SEVERITY_LABELS.remote_warning}s`,
   },
   {
     severity: "in_person_warning",
     label: `Add ${INCIDENT_SEVERITY_LABELS.in_person_warning.toLowerCase()}`,
-    hoverLabel: `${INCIDENT_SEVERITY_LABELS.in_person_warning}s`,
   },
   {
     severity: "citation",
     label: `Add ${INCIDENT_SEVERITY_LABELS.citation.toLowerCase()}`,
-    hoverLabel: `${INCIDENT_SEVERITY_LABELS.citation}s`,
   },
 ];
 
@@ -96,7 +98,7 @@ const PartyCard = memo(function PartyCard({
       };
 
   const articleClass = cn(
-    "border-b border-border px-4 py-2 last:border-b-0 transition-colors",
+    "border-b border-border px-4 py-2 transition-colors",
     onClick
       ? isActive
         ? "cursor-pointer bg-primary/10 ring-1 ring-primary/20"
@@ -111,22 +113,46 @@ const PartyCard = memo(function PartyCard({
       onClick={onClick}
       className={articleClass}
     >
-      <div className="space-y-1">
-        {/* Header: date/time + address + menu */}
+      <div className="pt-1">
+        {/* Header: address + date/time + menu */}
         <div className="flex items-start justify-between gap-3">
           <div>
-            {data.hasParty && (
-              <p className="content-bold font-bold text-secondary">
+            <div className="flex gap-1.5">
+              <p className="content-bold">
+                {data.hasParty
+                  ? formatAddress(data.party.location, [
+                      "street_number",
+                      "street_name",
+                      "unit",
+                      "city",
+                      "zip_code",
+                    ])
+                  : data.formattedAddress}
+              </p>
+              <a
+                href={buildMapsUrl(
+                  data.hasParty
+                    ? data.party.location.formatted_address
+                    : data.formattedAddress,
+                  data.hasParty
+                    ? data.party.location.google_place_id
+                    : data.locationPlaceId
+                )}
+                target="_blank"
+                rel="noopener noreferrer"
+                onClick={(e) => e.stopPropagation()}
+                className="flex text-muted-foreground hover:text-foreground transition-colors shrink-0 translate-y-px"
+                aria-label="Open in Google Maps"
+              >
+                <ExternalLink className="size-4" strokeWidth={2.5} />
+              </a>
+            </div>
+            {data.hasParty ? (
+              <p className="content-sub">
                 {format(data.party.party_datetime, "M/d/yyyy")} @{" "}
                 {formatTime(data.party.party_datetime)}
               </p>
-            )}
-            <p className="content-bold font-bold text-secondary">
-              {data.hasParty
-                ? data.party.location.formatted_address
-                : data.formattedAddress}
-            </p>
-            {!data.hasParty && (
+            ) : (
               <p className="content text-muted-foreground text-sm italic">
                 No party registered at this location
               </p>
@@ -163,70 +189,50 @@ const PartyCard = memo(function PartyCard({
 
         {/* Contacts with dotted separator */}
         {data.hasParty && (
-          <div className="grid grid-rows-2 gap-y-2 mb-2 mr-4">
-            <div className="flex flex-row justify-between gap-x-0.5">
-              <p className="content text-secondary self-end -mb-1.5">
-                {data.party.contact_one.first_name}{" "}
-                {data.party.contact_one.last_name}
-              </p>
-              <span className="flex-1 h-0.5 bg-[radial-gradient(circle,currentColor_1px,transparent_1px)] bg-size-[6px_2px] bg-repeat-x self-end text-secondary" />
-              <div className="flex flex-row gap-x-1 -mb-1.5">
-                <PhoneLink
-                  phoneNumber={data.party.contact_one.phone_number ?? "—"}
-                  contactPreference={data.party.contact_one.contact_preference}
-                  onClick={(e) => e.stopPropagation()}
-                />
-                <p className="content text-secondary">{" - "}</p>
-                <p className="content text-secondary">
-                  {formatContactPreference(
-                    data.party.contact_one.contact_preference
-                  )}
-                </p>
-              </div>
-            </div>
-            <div className="flex flex-row justify-between gap-x-0.5">
-              <p className="content text-secondary self-end -mb-1.5">
-                {data.party.contact_two.first_name}{" "}
-                {data.party.contact_two.last_name}
-              </p>
-              <span className="flex-1 h-0.5 bg-[radial-gradient(circle,currentColor_1px,transparent_1px)] bg-size-[6px_2px] bg-repeat-x self-end text-secondary" />
-              <div className="flex flex-row gap-x-1 -mb-1.5">
-                <PhoneLink
-                  phoneNumber={data.party.contact_two.phone_number}
-                  contactPreference={data.party.contact_two.contact_preference}
-                  onClick={(e) => e.stopPropagation()}
-                />
-                <p className="content text-secondary">{" - "}</p>
-                <p className="content text-secondary">
-                  {formatContactPreference(
-                    data.party.contact_two.contact_preference
-                  )}
-                </p>
-              </div>
-            </div>
+          <div className="flex flex-col py-1 mr-4">
+            {[data.party.contact_one, data.party.contact_two].map(
+              (contact, idx) => (
+                <div
+                  key={idx}
+                  className="flex flex-row items-baseline justify-between gap-x-0.5"
+                >
+                  <p className="content text-secondary">
+                    {contact.first_name} {contact.last_name}
+                  </p>
+                  <span className="flex-1 h-0.5 bg-[radial-gradient(circle,currentColor_1px,transparent_1px)] bg-size-[6px_2px] bg-repeat-x text-secondary" />
+                  <div className="flex flex-row items-baseline gap-x-1">
+                    <PhoneLink
+                      phoneNumber={contact.phone_number ?? "—"}
+                      contactPreference={contact.contact_preference}
+                      onClick={(e) => e.stopPropagation()}
+                    />
+                    <p className="content text-secondary">{" - "}</p>
+                    <p className="content text-secondary">
+                      {formatContactPreference(contact.contact_preference)}
+                    </p>
+                  </div>
+                </div>
+              )
+            )}
           </div>
         )}
 
         {/* Flags + hold expiration */}
-        <div className="flex items-center justify-between mt-3">
+        <div className="flex items-center justify-between">
           <div className="flex items-center gap-3">
-            {INCIDENT_MENU_ITEMS.map(({ severity, hoverLabel }) => (
-              <HoverCard key={severity} openDelay={0} closeDelay={4}>
-                <HoverCardTrigger asChild>
-                  <div className="flex items-center gap-1 content-bold font-bold text-foreground">
-                    {countBySeverity[severity]}
-                    <IncidentFlag type={severity} />
-                  </div>
-                </HoverCardTrigger>
-                <HoverCardContent>
-                  <p>{hoverLabel}</p>
-                </HoverCardContent>
-              </HoverCard>
+            {INCIDENT_MENU_ITEMS.map(({ severity }) => (
+              <div
+                key={severity}
+                className="flex items-center gap-1 content-bold font-bold text-foreground"
+              >
+                {countBySeverity[severity]}
+                <IncidentFlag type={severity} hoverCard />
+              </div>
             ))}
           </div>
           {data.hasParty && data.party.location.hold_expiration && (
             <div className="flex flex-row gap-2 justify-end items-center mr-4">
-              <HoverCard openDelay={0} closeDelay={4}>
+              <HoverCard>
                 <HoverCardTrigger asChild>
                   <AlertTriangle size="18px" className="text-destructive" />
                 </HoverCardTrigger>
