+++ frontend/src/lib/auth/route-access.ts
@@ -8,20 +8,20 @@ export const ALLOWED_ROLES_FOR_PATH: Record<string, AppRole[]> = {
   "/new-party": STUDENT_AREA_ROLES,
   "/profile": STUDENT_AREA_ROLES,
   "/staff": ["staff", "admin"],
-  "/police/admin": ["police_admin"],
+  "/police/admin": ["police_admin", "admin"],
   "/police": ["officer", "police_admin", "admin"],
 };
 
-export function getRequiredRolesForPath(pathname: string): AppRole[] | null {
+export function getAllowedRoles(pathname: string): AppRole[] {
   if (pathname === "/") return STUDENT_AREA_ROLES;
   for (const [prefix, roles] of Object.entries(ALLOWED_ROLES_FOR_PATH)) {
     if (pathname === prefix || pathname.startsWith(`${prefix}/`)) return roles;
   }
-  return null;
+  return [];
 }
 
 export function isStudentAreaPath(pathname: string): boolean {
-  return getRequiredRolesForPath(pathname)?.includes("student") ?? false;
+  return getAllowedRoles(pathname).includes("student");
 }
 
 export function getDashboardPath(role: AppRole | undefined): string {
